# spike/phase5-rest/sample-output/META-INF/openapi/openapi-v1.json
{
  "openapi" : "3.0.3",
  "info" : {
    "title" : "AI-ATLAS Generated API",
    "description" : "Auto-generated API from @AgenticExposed services",
    "version" : "1.0.0"
  },
  "paths" : {
    "/api/v1/customers" : {
      "get" : {
        "summary" : "Customers",
        "operationId" : "findAll",
        "responses" : {
          "200" : {
            "description" : "Success",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/CustomerDto"
                  }
                }
              }
            }
          }
        }
      },
      "post" : {
        "summary" : "Customers",
        "operationId" : "create",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/CustomerDto"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "201" : {
            "description" : "Created",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CustomerDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/customers/{id}" : {
      "get" : {
        "summary" : "Customers",
        "operationId" : "findById",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Success",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CustomerDto"
                }
              }
            }
          }
        }
      },
      "put" : {
        "summary" : "Customers",
        "operationId" : "update",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/CustomerDto"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "Success",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CustomerDto"
                }
              }
            }
          }
        }
      },
      "delete" : {
        "summary" : "Customers",
        "operationId" : "deleteById",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "204" : {
            "description" : "No Content"
          }
        }
      }
    },
    "/api/v1/customers/by-name/{name}" : {
      "get" : {
        "summary" : "Customer by name",
        "operationId" : "findByName",
        "parameters" : [ {
          "name" : "name",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Success",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CustomerDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/customers/activate" : {
      "post" : {
        "summary" : "Customers",
        "operationId" : "activate",
        "parameters" : [ {
          "name" : "id",
          "in" : "query",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Success",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CustomerDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/orders/{id}" : {
      "get" : {
        "summary" : "Order by id",
        "operationId" : "get",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Success",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/OrderDto"
                }
              }
            }
          }
        }
      },
      "delete" : {
        "summary" : "Cancel an order",
        "operationId" : "cancel",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "204" : {
            "description" : "No Content"
          }
        }
      }
    },
    "/api/v1/orders" : {
      "get" : {
        "summary" : "Orders by status",
        "operationId" : "byStatus",
        "parameters" : [ {
          "name" : "status",
          "in" : "query",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Success",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/OrderDto"
                  }
                }
              }
            }
          }
        }
      },
      "post" : {
        "summary" : "Place an order",
        "operationId" : "place",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/OrderDto"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "201" : {
            "description" : "Created",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/OrderDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/orders/{id}/status" : {
      "patch" : {
        "summary" : "Change an order's status",
        "operationId" : "changeStatus",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        }, {
          "name" : "status",
          "in" : "query",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Success",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/OrderDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/orders/count" : {
      "get" : {
        "summary" : "Number of orders",
        "operationId" : "count",
        "responses" : {
          "200" : {
            "description" : "Success",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "integer",
                  "format" : "int64"
                }
              }
            }
          }
        }
      }
    }
  },
  "components" : {
    "schemas" : {
      "OrderDto" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "integer",
            "description" : "Order id",
            "format" : "int64"
          },
          "status" : {
            "type" : "string",
            "description" : "Status"
          }
        },
        "description" : "An order"
      },
      "CustomerDto" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "integer",
            "description" : "Customer id",
            "format" : "int64"
          },
          "name" : {
            "type" : "string",
            "description" : "Name"
          }
        },
        "description" : "A customer"
      }
    }
  }
}